+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/ConfigurationTest.php
@@ -801,6 +801,7 @@ class ConfigurationTest extends TestCase
             ],
             'property_info' => [
                 'enabled' => !class_exists(FullStack::class),
+                'with_constructor_extractor' => false,
             ],
             'router' => [
                 'enabled' => false,